fix up class tests
Removed over 10 other tests, and moved many to a central make_bid_test - in the hopes of removing more later.

Added in the needed tests for the make class changes with first/last name
